system-stub/requirements.txt

annotated-types==0.7.0
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.7
paho-mqtt==1.6.1
pillow==10.4.0
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.0.7
numpy==1.26.3
